add use of api ca file if provided for statsd group check
