+++ internal/config/ssl.go
@@ -7,8 +7,6 @@ package config
 
 import (
 	"net"
-	"os"
-	"path/filepath"
 
 	"github.com/pkg/errors"
 	"github.com/spf13/viper"
@@ -41,44 +39,4 @@ func validateSSLOptions() error {
 	viper.Set(KeySSLKeyFile, key)
 
 	return nil
-}
-
-func verifyFile(fileName string) (string, error) {
-	if fileName == "" {
-		return "", errors.New("Invalid file name (empty)")
-	}
-
-	var absFileName string
-	var fi os.FileInfo
-	var err error
-
-	absFileName, err = filepath.Abs(fileName)
-	if err != nil {
-		return "", err
-	}
-
-	fileName = absFileName
-
-	fi, err = os.Stat(fileName)
-	if os.IsNotExist(err) {
-		return "", err
-	}
-
-	if os.IsPermission(err) {
-		return "", err
-	}
-
-	if !fi.Mode().IsRegular() {
-		return "", errors.New("not a regular file")
-	}
-
-	// also try opening, to verify permissions
-	// if last directory on path is not accessible to user, stat doesn't return EPERM
-	f, err := os.Open(fileName)
-	if err != nil {
-		return "", err
-	}
-	f.Close()
-
-	return fileName, nil
 }
